// src/routes/tasks/data/schema.ts
import { z } from "zod";

// We're keeping a simple non-relational schema here.
// IRL, you will have a schema for your data models.
export const taskSchema = z.object({
  id: z.string(),
  title: z.string(),
  notes: z.string(),
  status: z.enum(["draft", "progress", "done", "cancel"]),
  label: z.enum(["bug", "feature", "documentation", "enhancement"]),
  priority: z.enum(["high", "medium", "low"]),
  createdAt: z.string(),
  updatedAt: z.string().nullable(),
  startAt: z.string().nullable(),
  completedAt: z.string().nullable(),
});

export type Task = z.infer<typeof taskSchema>;